requirements.txt
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.2.0
scikit-image==0.22.0
rembg==2.0.65
python-magic==0.4.27
tqdm==4.66.1
onnxruntime==1.17.0
fastapi==0.109.2
python-multipart==0.0.9
uvicorn==0.27.1
httpx==0.26.0 